server/requirements.txt
flask
flask-cors
google-cloud-bigquery
pandas
gunicorn
pyarrow